feedback implemented

+++ Week3/homework/Ex2Transactions/transactions-create-tables.js
@@ -11,7 +11,7 @@ export const TABLES = [
         query: `CREATE TABLE IF NOT EXISTS account_changes(
         change_number INT AUTO_INCREMENT PRIMARY KEY,
         account_number INT,
-        amount FLOAT,
+        amount DECIMAL,
         changed_date DATE, 
         remark VARCHAR(200),
         FOREIGN KEY (account_number) REFERENCES account(account_number)
